requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
blinker==1.3
Brlapi==0.6.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
defer==1.0.6
feedfinder2==0.0.4
feedparser==5.2.1
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jieba3k==0.35.1
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==4.1.1
Mako==1.0.3
MarkupSafe==0.23
newspaper3k==0.2.6
nltk==3.2.5
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
parsel==1.4.0
pexpect==4.0.1
Pillow==5.0.0
plainbox==0.25
ptyprocess==0.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
PyDispatcher==2.0.5
pygobject==3.20.0
PyJWT==1.3.0
pyOpenSSL==17.5.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
pyxdg==0.25
PyYAML==3.12
queuelib==1.4.2
reportlab==3.3.0
requests==2.18.4
requests-file==1.4.3
Scrapy==1.5.0
service-identity==17.0.0
sessioninstaller==0.0.0
six==1.11.0
system-service==0.3
tldextract==2.2.0
Twisted==17.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.0
virtualenv==15.1.0
w3lib==1.19.0
xdiagnose==3.8.4
xkit==0.0.0
XlsxWriter==0.7.3
zope.interface==4.4.3
